+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+* **API** [fieldFor][doc:BitStructure.fieldFor].
 * Fix incorrect encoding of [boolean-valued bit fields][doc:Boolean].
 * **API** Add [offsetOf][doc:Structure.offsetOf] and [layoutFor][doc:Structure.layoutFor].